# Optimize `String#parameterize`