@Earlopain 2025-03-04
Alright, fixed the tests and kept most of the speed. It's now only slower when the separator is more than one character long (but still faster than on main). I updated the benchmark script and results in the description above.